refactor module contained to be tied to being added and removed to save event ordering
