+++ Content.Shared/Modules/Components/Modules/ModuleContainedComponent.cs
@@ -6,5 +6,5 @@ namespace Content.Shared.Modules.Components.Modules;
 public sealed partial class ModuleContainedComponent : Component
 {
     [DataField, AutoNetworkedField]
-    public EntityUid? Container;
+    public EntityUid Container;
 }
