+++ Content.Shared/Modules/Modules/ModSuitDeployableContainerVisualsSystem.cs
@@ -6,7 +6,7 @@ namespace Content.Shared.Modules.Modules;
 
 public sealed partial class ModSuitDeployableContainerVisualsSystem : EntitySystem
 {
-    [Dependency] private readonly ModuleContainedSystem _moduleContained = default!;
+    [Dependency] private readonly SharedModuleSystem _module = default!;
     [Dependency] private readonly SharedModSuitSystem _modSuit = default!;
 
     public override void Initialize()
@@ -18,7 +18,7 @@ public sealed partial class ModSuitDeployableContainerVisualsSystem : EntitySyst
 
     private void OnGetVisualEntity(Entity<ModSuitDeployableContainerVisualsComponent> ent, ref ModuleContainerVisualsGetVisualEntityEvent args)
     {
-        if (!_moduleContained.TryGetContainer(ent.Owner, out var container))
+        if (!_module.TryGetContainer(ent.Owner, out var container))
             return;
 
         if (_modSuit.TryGetDeployedPart(container.Value, ent.Comp.PartType, out var foundPart))
