+++ Content.Shared/Modules/Modules/ToggleableModuleSystem.cs
@@ -9,7 +9,7 @@ namespace Content.Shared.Modules.Modules;
 public sealed partial class ToggleableModuleSystem : EntitySystem
 {
     [Dependency] private readonly SharedPopupSystem _popup = default!;
-    [Dependency] private readonly ModuleContainedSystem _moduleContained = default!;
+    [Dependency] private readonly SharedModuleSystem _module = default!;
     [Dependency] private readonly IGameTiming _timing = default!;
 
     public override void Initialize()
@@ -23,9 +23,7 @@ public sealed partial class ToggleableModuleSystem : EntitySystem
 
     private void OnRemoved(Entity<ToggleableModuleComponent> ent, ref ModuleRemovedContainerEvent args)
     {
-        ent.Comp.Toggled = false;
-        Dirty(ent);
-        RaiseToggleEvents(ent, null);
+        RaiseToggleEvents(ent!, false, null);
     }
 
     private void OnEnabled(Entity<ToggleableModuleComponent> ent, ref ModuleEnabledEvent args)
@@ -67,7 +65,7 @@ public sealed partial class ToggleableModuleSystem : EntitySystem
         if (ent.Comp.Toggled == toggle)
             return;
 
-        if (!_moduleContained.TryGetContainer(ent.Owner, out var container))
+        if (!_module.TryGetContainer(ent.Owner, out var container))
             return;
 
         var beforeEv = new ModuleToggleAttemptEvent(toggle, container.Value, user);
@@ -86,14 +84,7 @@ public sealed partial class ToggleableModuleSystem : EntitySystem
             return;
         }
 
-        ent.Comp.Toggled = toggle;
-        Dirty(ent);
-        RaiseToggleEvents((ent.Owner, ent.Comp), user);
-    }
-
-    private void RaiseToggleEvents(Entity<ToggleableModuleComponent> ent, EntityUid? user)
-    {
-        RaiseToggleEvents(ent.AsNullable(), ent.Comp.Toggled, user);
+        RaiseToggleEvents(ent, toggle, null);
     }
 
     [PublicAPI]
@@ -102,7 +93,7 @@ public sealed partial class ToggleableModuleSystem : EntitySystem
         if (!Resolve(ent.Owner, ref ent.Comp))
             return;
 
-        if (!_moduleContained.TryGetContainer(ent.Owner, out var container))
+        if (!_module.TryGetContainer(ent.Owner, out var container))
             return;
 
         if (toggled)
