+++ Content.Shared/Modules/SharedModuleSystem.cs
@@ -1,11 +1,13 @@
+using System.Diagnostics.CodeAnalysis;
 using Content.Shared.Hands.EntitySystems;
 using Content.Shared.Modules.Components;
 using Content.Shared.Modules.Components.Modules;
 using Content.Shared.Modules.Events;
 using Content.Shared.Modules.ModSuit;
+using Content.Shared.Modules.ModSuit.Components;
 using Content.Shared.Modules.ModSuit.Events;
 using Content.Shared.Modules.ModSuit.UI;
-using Content.Shared.Modules.Modules;
+using Content.Shared.Modules.ModSuit.UI.Modules;
 using Content.Shared.Popups;
 using Content.Shared.Whitelist;
 using JetBrains.Annotations;
@@ -19,7 +21,6 @@ public abstract partial class SharedModuleSystem : EntitySystem
     [Dependency] private readonly SharedContainerSystem _container = default!;
     [Dependency] private readonly SharedHandsSystem _hands = default!;
     [Dependency] private readonly SharedModSuitSystem _modSuit = default!;
-    [Dependency] private readonly ModuleContainedSystem _moduleContained = default!;
     [Dependency] private readonly SharedPopupSystem _popup = default!;
     [Dependency] private readonly INetManager _net = default!;
     [Dependency] private readonly EntityWhitelistSystem _whitelist = default!;
@@ -42,7 +43,8 @@ public abstract partial class SharedModuleSystem : EntitySystem
         SubscribeLocalEvent<ModuleContainerComponent, EntInsertedIntoContainerMessage>(OnContainerInserted);
         SubscribeLocalEvent<ModuleContainerComponent, EntRemovedFromContainerMessage>(OnContainerRemoved);
 
-        SubscribeLocalEvent<ModuleContainerComponent, ModSuitGetUiEntriesEvent>(OnGetModSuitUiEntry);
+        SubscribeLocalEvent<ModuleContainerComponent, ModSuitGetUiEntriesEvent>(OnContainerGetModSuitUiEntry);
+        SubscribeLocalEvent<ModuleContainedComponent, ModuleRelayedEvent<ModSuitGetUiEntriesEvent>>(OnModuleGetModSuitUiEntry);
 
         SubscribeAllEvent<ModSuitEjectButtonMessage>(OnModSuitEjectButton);
     }
@@ -137,6 +139,10 @@ public abstract partial class SharedModuleSystem : EntitySystem
         if (args.Container.ID != ent.Comp.ModuleContainerId)
             return;
 
+        var contained = EnsureComp<ModuleContainedComponent>(args.Entity);
+        contained.Container = ent.Owner;
+        Dirty(args.Entity, contained);
+
         var containerEv = new ModuleContainerModuleAddedEvent(args.Entity);
         RaiseLocalEvent(ent.Owner, containerEv);
 
@@ -154,9 +160,11 @@ public abstract partial class SharedModuleSystem : EntitySystem
 
         var moduleEv = new ModuleRemovedContainerEvent(ent.Owner);
         RaiseLocalEvent(args.Entity, moduleEv);
+
+        RemComp<ModuleContainedComponent>(args.Entity);
     }
 
-    private void OnGetModSuitUiEntry(Entity<ModuleContainerComponent> ent, ref ModSuitGetUiEntriesEvent args)
+    private void OnContainerGetModSuitUiEntry(Entity<ModuleContainerComponent> ent, ref ModSuitGetUiEntriesEvent args)
     {
         // add a blank one so we override incase there is no more modules
         foreach (var entry in args.Entries)
@@ -169,12 +177,53 @@ public abstract partial class SharedModuleSystem : EntitySystem
         RelayToModules(ent, ref args);
     }
 
+    private ModSuitBaseModuleBuiEntry GetDefaultModSuitBui(Entity<ModuleContainedComponent> ent)
+    {
+        return new ModSuitBaseModuleBuiEntry(CompOrNull<ModSuitModuleComplexityComponent>(ent.Owner)?.Complexity);
+    }
+
+    private void OnModuleGetModSuitUiEntry(Entity<ModuleContainedComponent> ent, ref ModuleRelayedEvent<ModSuitGetUiEntriesEvent> args)
+    {
+        ModSuitModuleBuiEntry? foundEntry = null;
+
+        foreach (var entry in args.Args.Entries)
+        {
+            if (entry is not ModSuitModuleBuiEntry moduleEntry)
+                continue;
+
+            foundEntry = moduleEntry;
+            break;
+        }
+
+        if (foundEntry == null)
+            return; // should never happen as the modsuit adds a blank one to reset any lingering modules
+
+        var ev = new ModSuitGetModuleUiEvent();
+        RaiseLocalEvent(ent.Owner, ev);
+
+        var netEntity = GetNetEntity(ent.Owner);
+        var buiEntry = GetDefaultModSuitBui(ent);
+
+        foreach (var entry in ev.BuiEntries)
+        {
+            if (entry.Priority < buiEntry.Priority)
+                continue;
+
+            buiEntry = entry;
+        }
+
+        var newModules = new KeyValuePair<NetEntity, ModSuitBaseModuleBuiEntry>[foundEntry.Modules.Length + 1];
+        Array.Copy(foundEntry.Modules, newModules, foundEntry.Modules.Length);
+        newModules[^1] = KeyValuePair.Create(netEntity, buiEntry);
+        foundEntry.Modules = newModules;
+    }
+
     private void OnModSuitEjectButton(ModSuitEjectButtonMessage args)
     {
         var module = GetEntity(args.Module);
         var container = GetEntity(args.Container);
 
-        if (_moduleContained.GetContainer(module) != container)
+        if (GetContainer(module) != container)
             return;
 
         if (!TryComp<ModuleContainerComponent>(container, out var moduleContainer))
@@ -204,7 +253,7 @@ public abstract partial class SharedModuleSystem : EntitySystem
     [PublicAPI]
     public void RaiseEventToContainer<T>(Entity<ModuleContainedComponent?> ent, T ev) where T : notnull
     {
-        if (_moduleContained.TryGetContainer(ent, out var container))
+        if (TryGetContainer(ent, out var container))
             RaiseLocalEvent(container.Value, ev);
     }
 
@@ -212,5 +261,41 @@ public abstract partial class SharedModuleSystem : EntitySystem
     public void UpdateUis(EntityUid uid)
     {
         _modSuit.UpdateUI(uid);
+    }
+
+    [PublicAPI]
+    public bool TryGetContainer(Entity<ModuleContainedComponent?> ent, [NotNullWhen(true)] out EntityUid? container)
+    {
+        container = null;
+
+        if (!Resolve(ent.Owner, ref ent.Comp))
+            return false;
+
+        container = ent.Comp.Container;
+        return container != null;
+    }
+
+    [PublicAPI]
+    public EntityUid? GetContainer(Entity<ModuleContainedComponent?> ent)
+    {
+        if (!Resolve(ent.Owner, ref ent.Comp, false))
+            return null;
+
+        return ent.Comp.Container;
+    }
+
+    [PublicAPI]
+    public bool TryGetUser(Entity<ModuleContainedComponent?> ent, [NotNullWhen(true)] out EntityUid? user)
+    {
+        user = null;
+
+        if (!Resolve(ent.Owner, ref ent.Comp))
+            return false;
+
+        var ev = new ModuleGetUserEvent();
+        RaiseLocalEvent(ent.Comp.Container, ev);
+
+        user = ev.User;
+        return ev.User != null;
     }
 }
