+++ Resources/Prototypes/Entities/Objects/Specific/Robotics/ModSuits/Modules/base.yml
@@ -3,7 +3,6 @@
   parent: BaseItem
   abstract: true
   components:
-  - type: ModuleContained
   - type: Item
   - type: Sprite
   - type: Tag
